pyproject.toml
[project]
name = "geriatric-history-bot"
version = "0.1.0"
description = "Voice AI system for geriatric patient intake"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "websockets>=15.0.1",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.9.0",
]
